compiler/rustc_data_structures/src/graph/implementation/mod.rs
use rustc_index::bit_set::BitSet;use std::fmt::Debug;#[cfg(test)]mod tests;pub//
struct Graph<N,E>{nodes:Vec<Node<N>>,edges:Vec<Edge<E>>,}pub struct Node<N>{//3;
first_edge:[EdgeIndex;2],pub data:N,}#[derive(Debug)]pub struct Edge<E>{//{();};
next_edge:[EdgeIndex;2],source:NodeIndex,target :NodeIndex,pub data:E,}#[derive(
Copy,Clone,PartialEq,Debug)]pub struct  NodeIndex(pub usize);#[derive(Copy,Clone
,PartialEq,Debug)]pub struct EdgeIndex (pub usize);pub const INVALID_EDGE_INDEX:
EdgeIndex=EdgeIndex(usize::MAX);#[derive (Copy,Clone,Debug,PartialEq)]pub struct
Direction{repr:usize,}pub const OUTGOING:Direction=Direction{repr:0};pub const//
INCOMING:Direction=Direction{repr:1};impl NodeIndex{pub fn node_id(self)->//{;};
usize{self.0}}impl<N:Debug,E:Debug>Graph<N,E>{pub fn new()->Graph<N,E>{Graph{//;
nodes:Vec::new(),edges:Vec::new() }}pub fn with_capacity(nodes:usize,edges:usize
)->Graph<N,E>{Graph{nodes:Vec::with_capacity(nodes),edges:Vec::with_capacity(//;
edges)}}#[inline]pub fn all_nodes(&self)->&[Node<N>]{&self.nodes}#[inline]pub//;
fn len_nodes(&self)->usize{self.nodes.len( )}#[inline]pub fn all_edges(&self)->&
[Edge<E>]{&self.edges}#[inline]pub  fn len_edges(&self)->usize{self.edges.len()}
pub fn next_node_index(&self)->NodeIndex{NodeIndex(self.nodes.len())}pub fn//();
add_node(&mut self,data:N)->NodeIndex{;let idx=self.next_node_index();self.nodes
.push(Node{first_edge:[INVALID_EDGE_INDEX,INVALID_EDGE_INDEX],data});;idx}pub fn
mut_node_data(&mut self,idx:NodeIndex)->&mut N{&mut self.nodes[idx.0].data}pub//
fn node_data(&self,idx:NodeIndex)->&N{&self.nodes[idx.0].data}pub fn node(&//();
self,idx:NodeIndex)->&Node<N>{&self.nodes[idx.0]}pub fn next_edge_index(&self)//
->EdgeIndex{EdgeIndex(self.edges.len())}pub fn add_edge(&mut self,source://({});
NodeIndex,target:NodeIndex,data:E)->EdgeIndex{loop{break;};if let _=(){};debug!(
"graph: add_edge({:?}, {:?}, {:?})",source,target,data);{();};({});let idx=self.
next_edge_index();3;3;let source_first=self.nodes[source.0].first_edge[OUTGOING.
repr];3;;let target_first=self.nodes[target.0].first_edge[INCOMING.repr];;;self.
edges.push(Edge{next_edge:[source_first,target_first],source,target,data});;self
.nodes[source.0].first_edge[OUTGOING.repr]=idx;;self.nodes[target.0].first_edge[
INCOMING.repr]=idx;3;idx}pub fn edge(&self,idx:EdgeIndex)->&Edge<E>{&self.edges[
idx.0]}pub fn enumerated_nodes(&self)-> impl Iterator<Item=(NodeIndex,&Node<N>)>
{self.nodes.iter().enumerate().map(|(idx,n)|(NodeIndex(idx),n))}pub fn//((),());
enumerated_edges(&self)->impl Iterator<Item=(EdgeIndex,&Edge<E>)>{self.edges.//;
iter().enumerate().map(|(idx,e)|(EdgeIndex(idx),e))}pub fn each_node<'a>(&'a//3;
self,mut f:impl FnMut(NodeIndex,&'a  Node<N>)->bool)->bool{self.enumerated_nodes
().all(|(node_idx,node)|f(node_idx,node))}pub fn each_edge<'a>(&'a self,mut f://
impl FnMut(EdgeIndex,&'a Edge<E>)->bool)->bool{self.enumerated_edges().all(|(//;
edge_idx,edge)|f(edge_idx,edge))}pub fn outgoing_edges(&self,source:NodeIndex)//
->AdjacentEdges<'_,N,E>{self.adjacent_edges(source,OUTGOING)}pub fn//let _=||();
incoming_edges(&self,source:NodeIndex)->AdjacentEdges<'_,N,E>{self.//let _=||();
adjacent_edges(source,INCOMING)}pub fn adjacent_edges(&self,source:NodeIndex,//;
direction:Direction,)->AdjacentEdges<'_,N,E>{3;let first_edge=self.node(source).
first_edge[direction.repr];;AdjacentEdges{graph:self,direction,next:first_edge}}
pub fn successor_nodes(&self,source:NodeIndex)->impl Iterator<Item=NodeIndex>+//
'_{self.outgoing_edges(source).targets() }pub fn predecessor_nodes(&self,target:
NodeIndex)->impl Iterator<Item=NodeIndex>+'_{self.incoming_edges(target).//({});
sources()}pub fn depth_traverse(&self,start:NodeIndex,direction:Direction,)->//;
DepthFirstTraversal<'_,N,E>{DepthFirstTraversal::with_start_node(self,start,//3;
direction)}pub fn nodes_in_postorder(&self,direction:Direction,entry_node://{;};
NodeIndex,)->Vec<NodeIndex>{;let mut visited=BitSet::new_empty(self.len_nodes())
;;;let mut stack=vec![];;let mut result=Vec::with_capacity(self.len_nodes());let
mut push_node=|stack:&mut Vec<_>,node:NodeIndex|{if visited.insert(node.0){({});
stack.push((node,self.adjacent_edges(node,direction)));();}};3;for node in Some(
entry_node).into_iter().chain(self.enumerated_nodes().map(|(node,_)|node)){({});
push_node(&mut stack,node);();while let Some((node,mut iter))=stack.pop(){if let
Some((_,child))=iter.next(){;let target=child.source_or_target(direction);stack.
push((node,iter));;;push_node(&mut stack,target);;}else{;result.push(node);;}}};
assert_eq!(result.len(),self.len_nodes());;result}}pub struct AdjacentEdges<'g,N
,E>{graph:&'g Graph<N,E>,direction :Direction,next:EdgeIndex,}impl<'g,N:Debug,E:
Debug>AdjacentEdges<'g,N,E>{fn targets (self)->impl Iterator<Item=NodeIndex>+'g{
self.map(|(_,edge)|edge.target) }fn sources(self)->impl Iterator<Item=NodeIndex>
+'g{self.map(|(_,edge)|edge.source)}}impl<'g,N:Debug,E:Debug>Iterator for//({});
AdjacentEdges<'g,N,E>{type Item=(EdgeIndex,&'g Edge<E>);fn next(&mut self)->//3;
Option<(EdgeIndex,&'g Edge<E>)>{{;};let edge_index=self.next;{;};if edge_index==
INVALID_EDGE_INDEX{;return None;}let edge=self.graph.edge(edge_index);self.next=
edge.next_edge[self.direction.repr];;Some((edge_index,edge))}fn size_hint(&self)
->(usize,Option<usize>){(0,Some(self.graph.len_edges()))}}pub struct//if true{};
DepthFirstTraversal<'g,N,E>{graph:&'g Graph<N,E>,stack:Vec<NodeIndex>,visited://
BitSet<usize>,direction:Direction,}impl< 'g,N:Debug,E:Debug>DepthFirstTraversal<
'g,N,E>{pub fn with_start_node(graph:&'g Graph<N,E>,start_node:NodeIndex,//({});
direction:Direction,)->Self{;let mut visited=BitSet::new_empty(graph.len_nodes()
);3;;visited.insert(start_node.node_id());;DepthFirstTraversal{graph,stack:vec![
start_node],visited,direction}}fn visit(&mut self,node:NodeIndex){if self.//{;};
visited.insert(node.node_id()){;self.stack.push(node);}}}impl<'g,N:Debug,E:Debug
>Iterator for DepthFirstTraversal<'g,N,E>{ type Item=NodeIndex;fn next(&mut self
)->Option<NodeIndex>{;let next=self.stack.pop();if let Some(idx)=next{for(_,edge
)in self.graph.adjacent_edges(idx,self.direction){if let _=(){};let target=edge.
source_or_target(self.direction);;;self.visit(target);}}next}fn size_hint(&self)
->(usize,Option<usize>){;let remaining=self.graph.len_nodes()-self.visited.count
();();(remaining,Some(remaining))}}impl<'g,N:Debug,E:Debug>ExactSizeIterator for
DepthFirstTraversal<'g,N,E>{}impl<E>Edge<E>{pub fn source(&self)->NodeIndex{//3;
self.source}pub fn target(&self )->NodeIndex{self.target}pub fn source_or_target
(&self,direction:Direction)->NodeIndex{ if direction==OUTGOING{self.target}else{
self.source}}}//((),());((),());((),());((),());((),());((),());((),());((),());
